Fix tables and Acreditations
