+++ src/screens/Content/Content.jsx
@@ -21,7 +21,7 @@ import Posts from "./Posts/Posts";
 import CreatePosts from "./Posts/CreatePosts/CreatePosts";
 
 import Events from "./Events/Events";
-
+import CreateCertifications from "./Certifications/CreateCertifications/CreateCertifications";
 import Courses from "./Courses/Courses";
 import CreateCourses from "./Courses/CreateCourse/CreateCourses";
 
@@ -30,12 +30,11 @@ import {
 	PATH_CONTENT_DASHBOARD,
 	PATH_CREATE_PROJECTS,
 	PATH_PROJECTS,
-
+	PATH_CREATE_CERTIFICATION,
 	PATH_CERTIFICATIONS,
 	PATH_EVENTS,
 	PATH_POSTS,
 	PATH_CREATE_POSTS,
-
 	PATH_COURSES,
 	PATH_CREATE_COURSE
 } from "../../config/paths";
@@ -60,13 +59,13 @@ export default function ContentDashboard() {
 		keys.add(PATH_POSTS);
 		keys.add(PATH_CREATE_POSTS);
 		keys.add(PATH_CREATE_COURSE);
-
+		keys.add(PATH_CREATE_CERTIFICATION);
 		if (keys.has(pathname)) return;
 
 		navigate(PATH_PROJECTS); // Default
 	};
 
-	useEffect(checkIfNeedsToRedirect, [pathname]);
+	useEffect(checkIfNeedsToRedirect, [pathname, navigate]);
 
 	return (
 		<ContentContext.Provider value={{
@@ -89,6 +88,7 @@ export default function ContentDashboard() {
 					{pathname === PATH_POSTS && <Posts />}
 					{pathname === PATH_CREATE_POSTS && <CreatePosts />}
 					{pathname === PATH_CREATE_COURSE && <CreateCourses />}
+					{pathname === PATH_CREATE_CERTIFICATION && <CreateCertifications />}
 				</div>
 			</div>
 		</ContentContext.Provider>
